// src/api/types/operations-act-details-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Piattaforma Notifiche: API del BFF dei servizi RADD
 * API utilizzate dal Backend for Frontend dei servizi RADD.
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

// May contain unused imports in some cases
// @ts-ignore
import { OperationActDetailResponse } from './operation-act-detail-response';
// May contain unused imports in some cases
// @ts-ignore
import { OperationResponseStatus } from './operation-response-status';

/**
 * Id pratiche recuperate
 * @export
 * @interface OperationsActDetailsResponse
 */
export interface OperationsActDetailsResponse {
  /**
   *
   * @type {Array<OperationActDetailResponse>}
   * @memberof OperationsActDetailsResponse
   */
  elements?: Array<OperationActDetailResponse>;
  /**
   *
   * @type {boolean}
   * @memberof OperationsActDetailsResponse
   */
  result?: boolean;
  /**
   *
   * @type {OperationResponseStatus}
   * @memberof OperationsActDetailsResponse
   */
  status?: OperationResponseStatus;
}